+++ src/rer/customersatisfaction/browser/static/apps/customer_satisfaction_global/src/App/Comments/CommentsTable/Context/index.js
@@ -1,0 +1,10 @@
+import {createContext} from 'react';
+
+const FilterContext = createContext({});
+const FilterContextProvider = FilterContext.Provider;
+
+
+export {
+  FilterContextProvider,
+  FilterContext,
+};