+++ src/rer/customersatisfaction/browser/static/apps/customer_satisfaction_global/src/App/Comments/CommentsTable/utils/apiConfing.js
@@ -1,0 +1,12 @@
+// Constants
+
+const SITE_URL = '/';
+
+const SITES_LIST_URL = SITE_URL + 'sites-list';
+const COMMENTS_LIST_SUFFIX = '@customer-satisfaction';
+
+export {
+  SITE_URL,
+  SITES_LIST_URL,
+  COMMENTS_LIST_SUFFIX
+}