+++ src/rer/customersatisfaction/browser/static/apps/customer_satisfaction_global/src/App/Comments/index.js
@@ -1,0 +1,17 @@
+import Box from '@mui/material/Box';
+import CommentsTable from './CommentsTable'
+
+const Comments = () => {
+  return (
+    <Box sx={{
+      width: "98%",
+      mt: 5,
+      ml: "auto",
+      mr: "auto"
+    }}>
+      <CommentsTable />
+    </Box>
+  )
+}
+
+export default Comments;